==> 2_Transforming_htmls_to_dataframe.ipynb <==
from bs4 import BeautifulSoup
import os
from tqdm import tqdm
import pandas as pd

html_directory='Webpages/Aspiradoras'
htmls=os.listdir(html_directory)


def Get_information_from_amazon_html(soup):
    # Extract product title
    title = soup.find('span', id="productTitle")
    title_text = title.get_text(strip=True) if title else "Title not found"

    # Extract star rating
    stars_span = soup.find('span', id='acrPopover')
    stars_text = stars_span.get('title', 'Stars not found') if stars_span else "Stars not found"

    # Extract price
    price_whole = soup.find('span', class_="a-price-whole")
    price_fraction = soup.find('span', class_="a-price-fraction")
    price_symbol = soup.find('span', class_="a-price-symbol")

    if price_whole and price_fraction and price_symbol:
        price_text = f"{price_symbol.get_text(strip=True)}{price_whole.get_text(strip=True)}.{price_fraction.get_text(strip=True)}"
    else:
        price_text = "Price not found"

    # Extract general information located in productOverview_feature_div
    general_info_div = soup.find('div', id="productOverview_feature_div")
    if general_info_div:
        general_info_text = general_info_div.get_text(strip=True)
    else:
        general_info_text = "General information not found"

    # Extract "Acerca de este producto" section
    about_product_div = soup.find('div', id="featurebullets_feature_div")
    about_product = []
    if about_product_div:
        for item in about_product_div.select('ul.a-unordered-list li'):
            about_product.append(item.get_text(strip=True))
    else:
        about_product = ["About this product section not found"]

    # Extract technical details
    tech_details = {}
    for row in soup.select('#productDetails_techSpec_section_1 tr'):
        key = row.find('th').get_text(strip=True)
        value = row.find('td').get_text(strip=True)
        tech_details[key] = value

    information = {
        "Title": title_text,
        "Stars": stars_text,
        "Price": price_text,
        "General Information": general_info_text,
        "About This Product": about_product,
        "Technical Details": tech_details
    }
    return information


products=[]
for i,html in tqdm(enumerate(htmls)):
    try:
        with open(html_directory+"/"+html, 'r', encoding='utf-8') as file:
            soup = BeautifulSoup(file, 'html.parser')
            products.append(Get_information_from_amazon_html(soup))
            if products[-1]["Stars"]=="Stars not found":
                print(i,products[-1]["Stars"])
    except Exception as e:
        print(f"Error processing {html}: {e}")

df=pd.DataFrame(products)
df.to_csv("Databases/Parsed_htmls.csv", index=False)
